requirements.txt
# Core
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# RL / ML
gymnasium>=0.29.0
numpy>=1.24.0
stable-baselines3>=2.0.0

# OpenEnv
pydantic>=2.0.0
requests>=2.31.0

# Testing
pytest>=7.0.0
